revert: Framework splitting #9318
This reverts commit a96cafd4928ddc205e125a467d2f55338ed278fe.
