+++ .circleci/real_config.yml
@@ -30,13 +30,13 @@ parameters:
     default: ubuntu-2004:2024.01.1
   gpu-machine-image:
     type: string
-    default: linux-cuda-12:default
+    default: ubuntu-2004-cuda-11.2:202103-01
   # DEFAULT_PT_GPU_IMAGE: Pytorch training image reference used by the tests
   # Inject here as a parameter so that it is updated by bumpversion, and can
   # be referenced by --ee testing.
   default-pt-gpu-image:
     type: string
-    default: determinedai/pytorch-ngc-dev:8b3bea3
+    default: determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
   # Some python, go, and react dependencies are cached by circleci via `save_cache`/`restore_cache`.
   # If the dependencies stay the same, but the circleci code that would produce them is changed,
   # it may be necessary to invalidate the cache by incrementing this value.
@@ -251,7 +251,7 @@ commands:
       - when:
           condition: <<parameters.tf2>>
           steps:
-            - run: docker pull determinedai/pytorch-ngc-dev:8b3bea3
+            - run: docker pull determinedai/pytorch-tensorflow-cpu-dev:8b3bea3
 
   login-docker:
     parameters:
@@ -2320,9 +2320,9 @@ jobs:
       - store_test_results:
           path: /tmp/test-results
 
-  test-unit-harness-gpu-tf:
+  test-unit-harness-gpu:
     docker:
-      - image: determinedai/tensorflow-ngc-dev:8b3bea3
+      - image: determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
     resource_class: determined-ai/container-runner-gpu
     steps:
       - run: mkdir -p ~/.ssh && ssh-keyscan github.com >> ~/.ssh/known_hosts
@@ -2334,19 +2334,19 @@ jobs:
       - run: pip install mypy pytest coverage
       - install-codecov
       - setup-paths
-      - run: COVERAGE_FILE=/root/project/test-unit-harness-gpu-tf-pycov make -C harness test-gpu-tf
-      - run: coverage xml -i --data-file=./test-unit-harness-gpu-tf-pycov
+      - run: COVERAGE_FILE=/root/project/test-unit-harness-gpu-pycov make -C harness test-gpu
+      - run: coverage xml -i --data-file=./test-unit-harness-gpu-pycov
       - run: codecov -v -t $CODECOV_TOKEN -F harness
       - persist_to_workspace:
           root: .
           paths:
-            - test-unit-harness-gpu-tf-pycov
+            - test-unit-harness-gpu-pycov
       - store_test_results:
           path: /tmp/test-results
 
   test-unit-harness-pytorch2-gpu:
     docker:
-      - image: determinedai/pytorch-ngc-dev:8b3bea3
+      - image: determinedai/pytorch-cuda-dev:8b3bea3
     resource_class: determined-ai/container-runner-gpu
     steps:
       - run: mkdir -p ~/.ssh && ssh-keyscan github.com >> ~/.ssh/known_hosts
@@ -2370,7 +2370,7 @@ jobs:
 
   test-unit-harness-pytorch2-cpu:
     docker:
-      - image: determinedai/pytorch-ngc-dev:8b3bea3
+      - image: determinedai/pytorch-cpu-dev:8b3bea3
     steps:
       - run: mkdir -p ~/.ssh && ssh-keyscan github.com >> ~/.ssh/known_hosts
       - checkout
@@ -2393,7 +2393,7 @@ jobs:
 
   test-unit-harness-gpu-parallel:
     docker:
-      - image: determinedai/pytorch-ngc-dev:8b3bea3
+      - image: determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
     resource_class: determined-ai/container-runner-multi-gpu
     steps:
       - run: mkdir -p ~/.ssh && ssh-keyscan github.com >> ~/.ssh/known_hosts
@@ -2441,7 +2441,7 @@ jobs:
 
   test-unit-harness-tf2:
     docker:
-      - image: determinedai/tensorflow-ngc-dev:f17151a
+      - image: <<pipeline.parameters.docker-image>>
     steps:
       - checkout
       - add-and-fetch-upstream
@@ -2449,9 +2449,11 @@ jobs:
       - skip-if-only-github
       - skip-if-only-webui
       - install-codecov
-      - run: pip install mypy pytest coverage
-      - install-codecov
-      - setup-paths
+      - setup-python-venv:
+          install-python: false
+          determined: true
+          extra-requirements-file: "harness/tests/requirements/requirements-harness.txt"
+          executor: <<pipeline.parameters.docker-image>>
       - run: COVERAGE_FILE=$PWD/test-unit-harness-tf2-pycov make -C harness test-tf2
       - run: coverage xml -i --data-file=./test-unit-harness-tf2-pycov
       - run: codecov -v -t $CODECOV_TOKEN -F harness
@@ -3627,7 +3629,7 @@ jobs:
         type: string
         default: "1"
       environment-image:
-        default: determinedai/pytorch-ngc-dev:8b3bea3
+        default: determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
         type: string
       accel-node-taints:
         type: string
@@ -4027,7 +4029,7 @@ workflows:
       - test-unit-harness-pytorch2-cpu
       - test-unit-harness-pytorch2-gpu
 
-      - test-unit-harness-gpu-tf:
+      - test-unit-harness-gpu:
           filters: *any-upstream
       - test-unit-harness-gpu-deepspeed:
           filters: *any-upstream
@@ -4038,8 +4040,8 @@ workflows:
       - request-gpu-unit-tests:
           type: approval
           filters: *any-fork
-      - test-unit-harness-gpu-tf:
-          name: f-test-unit-harness-gpu-tf
+      - test-unit-harness-gpu:
+          name: f-test-unit-harness-gpu
           filters: *any-fork
       - test-unit-harness-gpu-deepspeed:
           name: f-test-unit-harness-gpu-deepspeed
@@ -4055,7 +4057,7 @@ workflows:
       - python-coverage:
           requires:
             - test-unit-harness-cpu
-            - test-unit-harness-gpu-tf
+            - test-unit-harness-gpu
             - test-unit-harness-gpu-parallel
             - test-unit-harness-tf2
             - test-unit-harness-pytorch2-cpu
