+++ docs/setup-cluster/slurm/slurm-requirements.rst
@@ -510,7 +510,7 @@ platform. There may be additional per-user configuration that is required.
 
    .. code:: bash
 
-      image=determinedai/pytorch-ngc-dev:8b3bea3
+      image=determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
       cd /shared/enroot/images
       enroot import docker://$image
       enroot create /shared/enroot/images/${image//[\/:]/\+}.sqsh
