+++ e2e_tests/tests/nightly/test_distributed.py
@@ -12,7 +12,7 @@ from tests import experiment as exp
 
 
 @pytest.mark.distributed
-@pytest.mark.parametrize("image_type", ["PT", "PT2"])
+@pytest.mark.parametrize("image_type", ["PT", "TF2", "PT2"])
 def test_mnist_pytorch_distributed(image_type: str) -> None:
     sess = api_utils.user_session()
     config = conf.load_config(conf.tutorials_path("mnist_pytorch/distributed.yaml"))
@@ -22,6 +22,8 @@ def test_mnist_pytorch_distributed(image_type: str) -> None:
         config = conf.set_pt_image(config)
     elif image_type == "PT2":
         config = conf.set_pt2_image(config)
+    elif image_type == "TF2":
+        config = conf.set_tf2_image(config)
     else:
         warnings.warn("Using default images", stacklevel=2)
