+++ examples/computer_vision/iris_tf_keras/adaptive.yaml
@@ -2,10 +2,6 @@ name: iris_tf_keras_adaptive_search
 data:
   train_url: http://download.tensorflow.org/data/iris_training.csv
   test_url: http://download.tensorflow.org/data/iris_test.csv
-environment:
-  image:
-    cpu: determinedai/tensorflow-ngc-dev:8b3bea3
-    gpu: determinedai/tensorflow-ngc-dev:8b3bea3
 hyperparameters:
   learning_rate:
     type: log
@@ -28,4 +24,4 @@ searcher:
   max_length:
     batches: 6400
   max_trials: 512
-entrypoint: python3 -m determined.launch.horovod --autohorovod --trial model_def:IrisTrial
+entrypoint: model_def:IrisTrial
