+++ examples/computer_vision/iris_tf_keras/distributed.yaml
@@ -2,10 +2,6 @@ name: iris_tf_keras_distributed
 data:
   train_url: http://download.tensorflow.org/data/iris_training.csv
   test_url: http://download.tensorflow.org/data/iris_test.csv
-environment:
-  image:
-    cpu: determinedai/tensorflow-ngc-dev:8b3bea3
-    gpu: determinedai/tensorflow-ngc-dev:8b3bea3
 hyperparameters:
   learning_rate: 1.0e-4
   learning_rate_decay: 1.0e-6
@@ -19,4 +15,4 @@ searcher:
   smaller_is_better: false
   max_length:
     batches: 2500
-entrypoint: python3 -m determined.launch.horovod --autohorovod --trial model_def:IrisTrial
+entrypoint: model_def:IrisTrial
