+++ examples/computer_vision/iris_tf_keras/iris_tf_keras_high_priority.yaml
@@ -2,10 +2,6 @@ name: iris_tf_keras_high_priority
 data:
   train_url: http://download.tensorflow.org/data/iris_training.csv
   test_url: http://download.tensorflow.org/data/iris_test.csv
-environment:
-  image:
-    cpu: determinedai/tensorflow-ngc-dev:8b3bea3
-    gpu: determinedai/tensorflow-ngc-dev:8b3bea3
 resources:
   slots_per_trial: 8
   resource_pool: defq_GPU_hipri
@@ -20,7 +16,7 @@ searcher:
   smaller_is_better: false
   max_length:
     batches: 500
-entrypoint: python3 -m determined.launch.horovod --autohorovod --trial model_def:IrisTrial
+entrypoint: model_def:IrisTrial
 min_validation_period:
   batches: 50
 min_checkpoint_period:
