+++ helm/charts/determined/values.yaml
@@ -31,8 +31,8 @@ defaultImages:
   kubeSchedulerPreemption: "determinedai/kube-scheduler:0.17.0"
 
   # default images for CPU and GPU environments
-  cpuImage: "determinedai/pytorch-ngc-dev:8b3bea3"
-  gpuImage: "determinedai/pytorch-ngc-dev:8b3bea3"
+  cpuImage: "determinedai/pytorch-tensorflow-cpu-dev:8b3bea3"
+  gpuImage: "determinedai/pytorch-tensorflow-cuda-dev:8b3bea3"
 
 # Install Determined enterprise edition.
 enterpriseEdition: false
