+++ master/pkg/schemas/expconf/const.go
@@ -8,7 +8,7 @@ const (
 
 // Default task environment docker image names.
 const (
-	CPUImage  = "determinedai/pytorch-ngc-dev:8b3bea3"
-	CUDAImage = "determinedai/pytorch-ngc-dev:8b3bea3"
+	CPUImage  = "determinedai/pytorch-tensorflow-cpu-dev:8b3bea3"
+	CUDAImage = "determinedai/pytorch-tensorflow-cuda-dev:8b3bea3"
 	ROCMImage = "determinedai/environments:rocm-5.0-pytorch-1.10-tf-2.7-rocm-622d512"
 )
