+++ schemas/test_cases/v0/experiment.yaml
@@ -47,8 +47,8 @@
       environment_variables: {}
       force_pull_image: false
       image:
-        cpu: determinedai/pytorch-ngc-dev:8b3bea3
-        cuda: determinedai/pytorch-ngc-dev:8b3bea3
+        cpu: determinedai/pytorch-tensorflow-cpu-dev:8b3bea3
+        cuda: determinedai/pytorch-tensorflow-cuda-dev:8b3bea3
         rocm: determinedai/environments:rocm-5.0-pytorch-1.10-tf-2.7-rocm-622d512
       pod_spec: null
       ports:
