+++ webui/react/src/fixtures/responses/trial-details/old-trial-config-noop-adaptive.json
@@ -30,8 +30,8 @@
   "name": "noop_adaptive",
   "environment": {
     "image": {
-      "cpu": "determinedai/pytorch-ngc-dev:8b3bea3",
-      "gpu": "determinedai/pytorch-ngc-dev:8b3bea3"
+      "cpu": "determinedai/pytorch-tensorflow-cpu-dev:8b3bea3",
+      "gpu": "determinedai/pytorch-tensorflow-cuda-dev:8b3bea3"
     },
     "ports": null,
     "force_pull_image": false,
